refactors the folder structure
